Testcase updated
