+++ package.json
@@ -3,6 +3,7 @@
   "type": "module",
   "devDependencies": {
     "@wdio/browser-runner": "^9.12.3",
+    "@wdio/cli": "^9.12.6",
     "@wdio/local-runner": "^9.12.3",
     "@wdio/mocha-framework": "^9.12.3",
     "@wdio/spec-reporter": "^9.12.3",
