--- test/pageobjects/Gajender.constants.js
@@ -1,4 +1,0 @@
-export const Credentials = {
-    username: 'standard_user',
-    passsword: 'secret_sauce'
-};
