--- test/pageobjects/Gajender.locators.js
@@ -1,38 +1,0 @@
-module.exports = {
-    loginPage : {
-        user_name:'#user-name',
-        password:'#password',
-        login_button:'#login-button',
-    },
-    homePage: {
-        inventory_item:'.inventory_item',
-        inventory_item_name:'.inventory_item_name',
-        inventory_item_desc:'.inventory_item_desc',
-        inventory_item_price:'.inventory_item_price',
-        add_to_cart_button:'.btn_inventory',
-        sortDropdown: '.product_sort_container',
-        productPrice: '.inventory_item_price',
-        productName: '.inventory_item_name',
-        productDescription: '.inventory_item_desc',
-    },
-    cartPage: {
-        productID1: '#add-to-cart-sauce-labs-backpack',
-        productID2: '#add-to-cart-sauce-labs-bike-light',
-        productID3: '#add-to-cart-sauce-labs-bolt-t-shirt',
-        cartTitle: '.title',
-        cartButton: '.shopping_cart_link',
-        productId_remove:'#remove-sauce-labs-bike-light',
-        cartBadge: '.shopping_cart_badge',
-        cartButton: '.shopping_cart_link',
-    },
-    checkoutPage: {
-        checkoutTitle: '.title',
-        checkoutButton: '#checkout',
-        first_name: '#first-name',
-        last_name: '#last-name',
-        postal_code: '#postal-code',
-        continue: '#continue',
-        finish: '#finish',
-        continue_shopping: '#continue-shopping',
-    }
-}
